+++ src/Nethermind/Nethermind.Consensus.AuRa/Contracts/BlockGasLimitContract.cs
@@ -18,11 +18,11 @@
 using System;
 using Nethermind.Abi;
 using Nethermind.Blockchain.Processing;
+using Nethermind.Consensus.AuRa.Json;
 using Nethermind.Core;
 using Nethermind.Core.Extensions;
 using Nethermind.Dirichlet.Numerics;
 using Nethermind.Evm;
-using Nethermind.Serialization.Json.Abi;
 
 namespace Nethermind.Consensus.AuRa.Contracts
 {
