+++ src/Nethermind/Nethermind.Consensus.AuRa/Contracts/RandomContract.cs
@@ -17,12 +17,12 @@
 using System;
 using System.Numerics;
 using Nethermind.Abi;
+using Nethermind.Consensus.AuRa.Json;
 using Nethermind.Core;
 using Nethermind.Core.Crypto;
 using Nethermind.Dirichlet.Numerics;
 using Nethermind.Evm;
 using Nethermind.Evm.Tracing;
-using Nethermind.Serialization.Json.Abi;
 
 namespace Nethermind.Consensus.AuRa.Contracts
 {
