+++ src/Nethermind/Nethermind.Consensus.AuRa/Contracts/ReportingValidatorContract.cs
@@ -15,9 +15,9 @@
 //  along with the Nethermind. If not, see <http://www.gnu.org/licenses/>.
 
 using Nethermind.Abi;
+using Nethermind.Consensus.AuRa.Json;
 using Nethermind.Core;
 using Nethermind.Evm;
-using Nethermind.Serialization.Json.Abi;
 using Nethermind.State;
 
 namespace Nethermind.Consensus.AuRa.Contracts
