+++ src/Nethermind/Nethermind.Consensus.AuRa/Contracts/TransactionPermissionContract.V3.cs
@@ -17,10 +17,10 @@
 
 using System;
 using Nethermind.Abi;
+using Nethermind.Consensus.AuRa.Json;
 using Nethermind.Core;
 using Nethermind.Core.Extensions;
 using Nethermind.Dirichlet.Numerics;
-using Nethermind.Serialization.Json.Abi;
 
 namespace Nethermind.Consensus.AuRa.Contracts
 {
