+++ src/Nethermind/Nethermind.Consensus.AuRa/Json/AbiParameterConverter.cs
@@ -22,7 +22,7 @@ using Nethermind.Abi;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
 
-namespace Nethermind.Serialization.Json.Abi
+namespace Nethermind.Consensus.AuRa.Json
 {
     internal static class AbiParameterConverterStatics
     {
